Add boolean to install nipm

+++ build.ps1
@@ -1,24 +1,24 @@
-Write-Output "Installing NIPM"
 $rootDirectory = (Get-Location).Path
 Write-Output "Current directory $rootDirectory"
 
-$nipmDownloadPath = 'http://download.ni.com/support/softlib/AST/NIPM/NIPackageManager18.0.2.exe'
-$nipmInstaller = Join-Path -Path $rootDirectory -ChildPath 'install-nipm.exe'
-
-Write-Output "Downloading NIPM from $nipmDownloadPath..."
-$webClient = New-Object System.Net.WebClient
-$webClient.DownloadFile($nipmDownloadPath, $nipmInstaller)
-Write-Output "...done"
-if (![System.IO.File]::Exists($nipmInstaller))
+$install_NIPM = $true
+if ($install_NIPM)
 {
-    throw "Could not find downloaded NIPM installer"
+    $nipmDownloadPath = 'http://download.ni.com/support/softlib/AST/NIPM/NIPackageManager18.0.2.exe'
+    $nipmInstaller = Join-Path -Path $rootDirectory -ChildPath 'install-nipm.exe'
+    Write-Output "Downloading NIPM from $nipmDownloadPath..."
+    $webClient = New-Object System.Net.WebClient
+    $webClient.DownloadFile($nipmDownloadPath, $nipmInstaller)
+    Write-Output "...done"
+    if (![System.IO.File]::Exists($nipmInstaller))
+    {
+        throw "Could not find downloaded NIPM installer"
+    }
+    Write-Output "Installing NIPM..."
+    Start-Process -FilePath $nipmInstaller -ArgumentList "/Q" -Wait
+    Write-Output "...done"
+    Remove-Item $nipmInstaller
 }
-
-Write-Output "Installing NIPM..."
-Start-Process -FilePath $nipmInstaller -ArgumentList "/Q" -Wait
-Write-Output "...done"
-
-Remove-Item $nipmInstaller
 
 $nipm = 'C:\Program Files\National Instruments\NI Package Manager\NIPackageManager.exe'
 if (![System.IO.File]::Exists($nipm))
